dapp: add PantherPoolV0 contract
Without this patch, there is no way to call the PoolV0 contract in the
dApp.

This is a problem because the contract is needed for the redemption of
the zAssets.

This patch solves the problem by adding environmental variables and
functions needed to use the PoolV0 contract in the dApp.

Issue-URL: https://app.clickup.com/t/26gtvqm