contracts: constant PRP_VIRTUAL_CONTRACT used
This commit accounts for changes in PrpGrantor made by the commit 4ee8c5f5.
Since now, the address of the "virtual token contract" for PRP ("virtual" since
the PRP is NOT a token) is defined by the PRP_VIRTUAL_CONTRACT rather than the
address of the `PantherPoolV0` contract (which may change).