dapp: truncation and formating of the private and public keys
Without this patch, both the public and private keys are truncated to
the BN254 field.

This is a problem because the private key lives in the scalar field of
the BabyJubJub suborder. So for any operations on the private key, we
would need to truncate the resulting integer in the scalar field of the
BabyJubJub. This is the field of size babyjubjub.suborder. The public
keys, however, are points on the BabyJubJub. In the affine form, each
point P is a pair (P.x,P.y) of two elements of this BN254 field. So any
time you add/double points on the BabyJubJub, you would need to
truncate the resulting x, y coordinates in the BN254 field.

This patch solves the problem by truncating the private key to the field
of size babyjubjub.suborder and the public key to the BN254 field. Also,
this patch adds assert functions to check that the private key is in the
field of size babyjubjub.suborder and public keys are within the BN254
field. Adds and fixes the tests.