Add a service-layer function for querying subgraph.
Without this patch, FE is not able to get data from the subgraph.

This is a problem because FE gets on-chain data from the subgraph instead of directly getting them from the network.

This patch solves the problem by adding a function to send the HTTP requests to the subgraph.
HTTP module has been added to create a customized Axios object. then the `customAxios` object is used to send the `post` request to `theGraph` endpoint.

Issue-URL: https://app.clickup.com/t/2e57d8x