dapp: remove leftLeafId from and add utxoData to the subgraph
Without this patch, the query contained a request for leftLeafId,
meanwhile, utxoData is missing.

This is a problem because the updated subgraph does not have leftLeafId
and as a result, returned the error. Moreover, the error was hidden in
the subgraph response with status 200. As a result, the query returned
the results of an empty array.

This patch solves the problem by removing leftLeafId from the query,
adding utxoData in the request, and checking if the subgraph has no
errors in the response.