Subgraph codes have been added.
Without this patch, There is no efficient way to get the user's leaf ids from the network.

This is a problem because we have some limitations in reading events from the polygon. this network does not let us query more than 1000 blocks. so as new blocks are mined, more time is iterations needed to retrieve the user's lead id and it increases the time for getting the final result.

This patch solves the problem by creating a subgraph to get data from the network.

Issue-URL: https://app.clickup.com/t/2dm2bv2