crypto: update the Merkle tree to work with more than 2048 quad leaves
Without this patch, the Merkle tree had a hardcoded limit of 2048 quad
leaves.

This is a problem because in the Pool contract Merkle trees are the size
of 65,536.

This patch solves the problem by updating leafIdToTreeIdAndTriadId() to
work with more than 2048 quad leaves and calculate the indices using
dynamic fowmula: maxQuadLeafId = 2 ** (treeDepth - 1) * 4. Also, this
patch adds logging of the commitment (leaf HEX) in generateMerkleProof().