dapp: early redemption of zZKP
Without this patch, users are not able to redeem their zZKP early.

We need to add support to the dApp for performing early redemption after
the advanced staking period ends, to allow users to redeem their zZKP
rewards in the case that the MASP v1 and core protocol is not launched
at that time.

This patch solves the problem by adding a new service function exit().
By calling this function, the user decodes UTXO data received from the
subgraph, deciphers the random secret, generates child spending keys,
checks if the nullifier is not spent, verifies that the commitment (leaf)
of the Merkle tree is the same as the commitment of the UTXO, generates
and checks Merkle path, and finally submits exit() transaction.

To redeem UTXO and perform all checks, the following data are required:
- leaf ID;
- creation time of UTXO;
- commitments (triad leaves);
- utxoData (ciphered text, the token amount, and address).

These additional data are put into the AdvancedStakeRewards slice and
are fetched from the Panther Pool subgraph.

Further, the AdvancedStakeRewards slice is updated to include the utxoIsSpent
field (boolean). This field is used to determine if the UTXO has been
spent based on the result returned after the call of poolContract.isSpent().
The parent exit function always returns the utxoIsSpent boolean.

Minor changes to support early redemption:
- updated pool contract ABI with verifyMerkleProof() to check the Merkle
path;
- updated react views.