dapp: lock redeem button until lockedTill date
Without this patch, a user can redeem UTXO even though the lock time is
not reached.

This is a problem because we are not allowing redeeming of UTXOs before
the lock time due to the commitment Merkle tree is not yet ready.

This patch solves the problem by locking the button in UI based on the
staking terms fetched from the contract.

Issue-URL: https://app.clickup.com/t/2fexxp4