dapp: fix rewards if no stakes in the subgraph
Without this patch, if no events are found for the address in the graph,
the returning value for rewards.staker will be null.

This is a problem because that throws an error when we are trying to
access rewards.staker.advancedStakingRewards during fetching of the
rewards from the subgraph.

This patch solves the problem by adding a check to ensure that the
rewards.staker is not null.