dapp: PRP and zZKP total rewards based on advancedStakesRewards slice
Without this patch, we calculate reward amounts in different places
based on the different sources. On the zAssets page, we calculate the
totals based on the AdvancedRewards slice fetched from the subgraph. On
the other hand, in the balance card, we calculate the rewards based on
the advanced stakes fetched from the Staking smart contract.

This is a problem because it may potentially result in mismatches of the
values. Also, the slices in redux states with almost the same names, and
duplication of the types/env variables produce confusion.

This patch solves the problem by calculating PRP and zZKP rewards based
on the advancedStakesRewards slice fetched from the subgraph.
Additionally, this change allows simplification of the
unclaimedStakesRewards slice to contain only classic stakes rewards. To
clarify its purpose, also rename it to totalUnclaimedClassicRewards.

Part of Issue-URL: https://app.clickup.com/t/2fez545