Add a missing `advancedActionHash()` to `lib/hash.ts`
Without this patch, there is no `advancedActionHash()` in `lib/hash.ts`.

This is a problem because it has been imported into another module and without it, running contracts locally throws an error.

This patch solves the problem by adding the missing function to `hash.ts`.
The `combinedHash(action: string, type: string)` is also added as a helper function to calculate the hash of `action` and `type` together and then returns the first 10 bytes of the hash (`0x` + 8 bytes). The `advancedActionHash()` and `classicActionHash()` functions calls this helper function to get the hash either for `ADVANCED` type or `CLASSIC` type.

Issue-URL: https://app.clickup.com/t/2dzeddq