dapp: persist advanced stakes rewards slice in local storage
Without this patch, on page reload of the assets page, all advanced
rewards (UTXOs) are shown including already redeemed (spent) ones.

This is a problem because it results in a bad UX as a user needs to
click redeem button and manually check if UTXO is spent or not on each
load of the page. At the same time, we cannot store the private keys in
the redux or local storage for easy check of UTXO status as it is
damaging the security.

This patch solves the problem by persisting advanced rewards redux slice
in local storage using the redux-persist package. In this way, during
page reload UTXO statuses (spent, unspent, undefined) are also populated
in the redux store. However, this persistence resulted in the other
issue, when the user switches accounts, rewards from the previous
account are also shown which was not correct. To overcome this issue,
this patch adds an address as a key to the redux slice of advanced
rewards, i.e. the slice is the map of address and rewards ID. Thus, the
selectors were updated to have additional parameters of address.
Although the straight addition of the address would undermine the
privacy as it would be saved in local storage, therefore, instead of the
address as the key, the Poseidon hash of the address (last 10 characters
of the hex form) is used in setter and getter methods.  Additionally,
returned boolean values of the exit() function were replaced with a more
accurate state UTXOStatus. Finally, react UI was updated to show only
unspent and undefined  UTXO statuses.

Issue-URL: https://app.clickup.com/t/26gtvqm