Add NotFound page for panther-core project
Without this patch, we don't have proper route handling for the undefined routes.

This is a problem because if we type any route on the browser manually (Ex: ./zAssets) and press Enter, then the current routing system will redirect user to the home page route (./ ), even if we type already defined routes such as ./faucet etc.

This patch solves the problem by changing line 100 in App.tsx file in a more appropriate way to load NotFoundPage if the user hits undefined routes on the browser instead of redirection to home page.

Issue-URL: https://app.clickup.com/t/2f1vz6q