dapp: update the PRP rewards calculation
Without this patch, we had an outdated math formula for PRP rewards
calculation. It was as follows: for each 1000 ZKP user gets 1 PRP.

This is a problem because, in the updated version of the smart contract,
the PRP reward is fixed and equals 10,000 PRP per stake.

This patch solves the problem by correctly calculating the PRP reward in
agreement with how it is calculated in the smart contract. Corresponding
tests are also updated.