dapp: refresh UTXOs statuses service function
Without this patch, all UTXOs cannot be refreshed to show accurate
status by signing the message only once.

This is a problem because it results in bad UX as the user needs to sign
the message for each UTXO to check the status.

This patch solves the problem by introducing the new service function:
refreshUTXOs(). This function asks for a signature only once and then
iterates over the list of UTXOs to check if the UTXO was spent or not.
The UTXO status check is done by a new supporting function
unpackUTXOAndDeriveKeys(). Among other variables, the function returns
the UTXOStatus. The same functionality was already present in the exit()
function and in order not to violate the DRY principle, it was extracted
to a new unpackUTXOAndDeriveKeys() function.

Issue-URL: https://app.clickup.com/t/2gq62tg