get exit time from pool v0 contract not from staking terms
Without this patch, the time when early zZKP redemption starts being
allowed was being taken from the lockedTill parameter in the staking
terms.  However early redemption has nothing to do with staking, and
the time it opens is determined by EXIT_TIME as set in the constructor
of the PantherPoolV0 contract.

So fix it to be taken from there.