dapp: refresh UTXO status in Redux
Without this patch, a user is not able to trigger refresh UTXO action.

This is a problem because this functionality is needed to reflect
accurately user balances.

This patch solves the problem by adding a new asyncThunk action in redux
that updates the statuses of the UTXOs using a newly created helper
service function. The UTXOs status changes are logged in console. Also,
the code adds additional TODO that will be dealt with in the follow-up
commits, namely, adding a new 'refresh utxos' button.

Issue-URL: https://app.clickup.com/t/2gwp7dm