dapp: save the lastRefreshTime time stamp of the latest refresh
Without this patch, there is no information about the last time the
UTXOs statuses were refreshed.

This is a problem because the balance of the user's wallet depends on
the status of the UTXOs (spent or not spent). If the UTXOs statuses are
not refreshed, the balance of the user wallet could be wrong. Showing
this information to the user may indicate to the user the accuracy of
the estimation of the user's asset balances.

This patch solves the problem by adding a timestamp to the
advancedRewards slice to store the last time the UTXOs statuses were
refreshed.

Issue-URL: https://app.clickup.com/t/2gwrun7