dapp: error handling on exit() call
Without this patch, errors during early redemption were not properly
handled. As a result, this led to masking underlying issues from the
user. For example, if a user declines signing of the message, the
notification with the error and the reason will not show up (it is
reported in the dev console solely).

This is a problem because the user will not know what went wrong and
why.

This patch solves the problem by properly handling the errors and
showing the notification during
1) when a user declines the signature of the message. If a user declines
   the signature, the error was returned as an object with a message key
   but was not an instance of Error. Therefore, on the catch block, the
   error was returned as a new Error. Due to changes in the
   deriveRootKeypairs() function, the returned values are checked if
   they are the instances of Error in all places where the function is
   called, including actions of the reducers. Also, the tests
   (keychain.test.ts) were updated to omit the error type of the
   returning value.
2) when the generation of the Merkle path throws an error. These errors
   were not properly caught and were reported only in the console.

Issue-URL: https://app.clickup.com/t/2ghbzay