dapp: hook up the last refresh time and refresh button to the UI
Without this patch, the latest refresh date is not shown, and the
refresh UTXO button will not work.

This patch solves the problem by hooking up existing UI elements to the
new functions. Also, the patch introduces new convenience
formatTimeSince() function to format time since xxx date, e.g. 4 minutes
ago. This function shows the time of the last sync. New selectors are
added to the AdvancedRewardsSlice to get the last refresh time and the
status which is used for the spinner. Additionally, the patch updated
the tooltip with the information about the purpose of the refresh
button.

Issue-URL: https://app.clickup.com/t/2hdyw96