dapp: add modal dialog on zAsset page load
Without this patch, when the zAsset page is loaded and there are UTXOs
in `UNDEFINED` status, then we automatically request the user to sign
the message to derive their root keys.

This is a problem because we don't explain why we do that, so this
will confuse the user.

This patch solves the problem by showing this modal dialog on the page
while the wallet is requesting/generating the signature and
automatically closing it once the signature is generated, or also if
the signature generation fails.

In order to track when to show this new modal dialog, we create a new
Redux slice `web3WalletLastAction` which tracks the most recently
executed action on the underlying web wallet.  This will also help us
protect the user against submitting a second web3 wallet action while
a first one is already in progress.

Issue-URL: https://app.clickup.com/t/2he21jk

ClickUp: #2he21jk

- [x] The commit messages all explain WHY the changes are needed.
- [x] The commit messages all explain HOW the changes solve the problem.
- [x] The commits are all strongly related.
- [x] None of the commits introduce any breakage.
- [x] None of the commits introduce any new technical debt.
- [ ] The commits update any related documentation.
- [ ] The commits include automated testing.
- [x] The commits have been thoroughly tested.