dapp: add script for setting Amplify environment variables
Without this patch, we are manually setting AWS Amplify environment
variables to mirror advanced staking timestamp parameters whose source
of truth is defined in smart contracts.

This is a problem because the environment variables can easily get out
of sync with the smart contracts by mistake.

This patch solves the problem by making read-only calls to the Staking
and PantherPoolV0 smart contracts to extract the relevant timestamps,
and then using the AWS CLI to update the Amplify app accordingly.

Issue-URL: https://app.clickup.com/t/2hkr2ch

ClickUp: #2hkr2ch