dapp: update text for when wallet is disconnected
When the wallet was disconnected, it was showing old text about
classic staking.  Since we don't know chainId until the wallet is
connected, we don't know whether to focus on classic
unstaking (e.g. on mainnet or matic) or advanced staking (mumbai), so
instead show a more generic message which encourages the user to
connect their wallet.

Issue-URL: https://app.clickup.com/t/2hkjgpr