dapp: add "open from" time to StakingInfo
The existing logic in the StakingInfo component within the staking tab
was checking whether staking was open according to the smart contract,
and if not, it was assuming that staking was closed, totally ignoring
the possibility that staking hadn't started yet.

So add the "open from" time to StakingInfo, so that this component can
be used even before staking is open.

A future commit will add a check to distinguish between staking being
open, and staking being open at some point in the future.

Issue-URL: https://app.clickup.com/t/2hkjgpr