dapp: switch formatTime from array function declaration
In general, using => array function declaration notation is more
appropriate either when the function body is just an expression, or
the function symbol needs to be scoped (e.g. as a closure within a
React functional component).

In contrast, declarations at the top level don't need this scoping.
And using `function` is more explicit, and in this case it's also more
consistent with other declarations.  So switch to this style.