dapp: display correct unstaking/redemption time on Welcome page
Without this patch, the time displayed for unstaking and redemption
on the Welcome page is hardcoded to a fixed value, which is obviously
incorrect.

This patch solves the problem by adding a new `ADVANCED_STAKING_T_UNLOCK`
environment variable and using that.

Note that the value of this should be identical to
`ADVANCED_STAKING_EARLY_REDEMPTION` which is currently unused but is
set by syncEnv.ts.

Issue-URL: https://app.clickup.com/t/2hdvjh3