contracts: allow testnet deployments to override EXIT_TIME
The exit time can not be updated after deployment, so if it passes, the contract needs to be deployed again. This makes testing hard because sometimes the ability/disability of redeeming rewards need to be tested quickly.

Making `EXIT_TIME` mutable does not seem to be appealing in the production code. So the `EXIT_TIME` has become private and the `view virtual` function named `exitTime()` has been added to read this value.
This trick helps to define a mutable `_testnetExitTime` in the `TestnetPantherPoolV0` and override the `exitTime()` to return the `_testnetExitTime`. There is also `updateExitTime()` to update this value.