contracts: ZAssetsRegistry refactored and made a separate contract
Before the commit, ZAssetsRegistry was a part of the PantherPoolV0 that inherited from the former.
It was a problem since:
(1) "separate concerns" were "mixed" in the PantherPoolV0 (UTXOs trees and core MASP transactions
on one hand, and the registry of supported zAssets, on the other hand)
(2) PantherPoolV0 bytecode was close to the 24kb limit, which complicated implementation of missing
core functionality on the coming upgrade to the V1.
On top of it, the registry did not support multiple "versions" of zAssets for different "scales" of
the same ERC-20 token.
Finally, amount scaling functions, which shall be extracted from the ZAssetRegistry and remain in
the PantherPoolV*, have been separated into the AmountConvertor smart contract.

This commit:
- Removes amount scaling functions from the ZAssetRegistry
- Transform the ZAssetRegistry into a stand-alone smart contract
- Introduce "versions" of zAssets with different scaling factors for an ERC-20 token
- Updates mock/fake contracts (used in tests) to support changes