contracts: PantherPoolV0 integrated w/ new PrpGrantor & ZAssetsRegistry
Before the commit, PrpGrantor and ZAssetsRegistry was a part of the PantherPoolV0 that inherited
from the (old versions of the) two formers. It was a problem since:
(1) "separate concerns" were "mixed" in the PantherPoolV0 (UTXOs trees and core MASP transactions
on one hand, the registry of supported zAssets and granting PRPs, on the other hand)
(2) PantherPoolV0 bytecode was close to the 24kb limit, which complicated implementation of missing
core functionality on the coming upgrade to the V1.
On top of it, the newer versions of the PrpGrantor and ZAssetsRegistry (also the AmountConvertor,
that contains scaling functions extracted from ZAssetsRegistry to leave them in the PantherPoolV*)
was introduced by recent commits.
This commit integrates stand-alone contracts (refactored), the PrpGrantor and the ZAssetRegistry,
with the PantherPoolV0.