dapp: move pool V0 exitTime retrieval into Redux slice
Without this patch, we call exitTime() on PantherPoolV0 once
for every row of the ZAsset table.

This is totally unnecessary since the value doesn't change, and it
causes spamming of the JSON-RPC endpoint, and slows down UI
performance.

This patch solves the problem by adding a new poolV0 slice and
caching the exitTime value in that.

Issue-URL: https://app.clickup.com/t/33kber3