dapp: improve errors when stakes are outside staking period
When using TestnetStaking, the staking terms can be updated, causing
existing stakes to become outside the new staking period.  In these
cases, calculating the zZKP rewards was logging errors to the console
which were slightly misleading and also a bit too alarmist.

So change from errors to warnings, and make the text more accurate and
informative by including the stake amount and the way the situation is
being handled.