dapp: store UTXOs in redux per chainId
Without this patch, when a user switches between the networks, the page
shows all his UTXOs as they are loaded from the persistent redux store.
The problem arises from the fact that we store the UTXOs as a map of
hashed addresses to the object of UTXOs in the persistent redux store.
This means that when the user switches between the networks, the page
shows the same UTXOs from the persistent redux store.

This is a problem because it shows incorrect information and will
mislead a user.

This patch solves the problem by storing UTXOs in the redux store per
chainId. Indeed, instead of a map of addressHash -> leafId ->
AdvancedStakeRewards, the patch implements a map of chainId ->
addressHash -> leafId -> AdvancedStakeRewards. To accomplish that, the
patch adds a chainId in selectors (getters), reducers (setters), and
thunks (actions) where needed. Additionally, all UI components are
updated to pass chainId into the selectors.

Issue-URL: https://app.clickup.com/t/314k8z0