dapp: fix incorrect PRP rewards value in the balance card
Without this patch, the value of PRP rewards shown on the balance card
is 0. This is because UI assumes all tokens (including PRP) are scaled
with 18 decimals. However, in the smart contract, the PRP rewards are
not scaled.

This is a problem because UI shows incorrect PRP rewards.

This patch solves the problem by adding an optional parameter to the
formatCurrency() function. A new scale parameter is used to specify the
number of decimals of the token. Additionally, this patch changes the
AddressBalances React component to use the new scale parameter.
Moreover, in the PrivateBalance and  PrivateZAssetRow React components
formatCurrency function with scale parameter replaced the artificial
addition of 18 decimals with utils.parseEther() function.

Issue-URL: https://app.clickup.com/t/363ag0n