dapp: total PRP includes PRP from the spent zZKP UTXOs
 Without this patch, the PRP from spent zZKP UTXOs are not included in
 the total PRP. As to our discord discussion
 (https://discord.com/channels/801604563115180063/973550711730217000/994180969089224744),
 during redemption of zZKP, a user does not necessarily lose PRPs.
 Therefore, we have decided to include PRPs rewards from rewards that
 have spent zZKP UTXOs in the calculation of the total PRP rewards
 earned by the user.

This patch solves the problem by adding an optional parameter to the
totalSelector of the advancedStakesRewards slice. This optional boolean
parameter "includeIfZZkpSpent" specifies whether to include PRPs from
the rewards that have spent zZKP UTXOs. If the optional parameter is not
specified, it defaults to false. Moreover, the patch updated
totalSelector to always discard UTXOStatus.UNDEFINED. Also, utxoStatus
in advancedStakesRewards slice was renamed to a more accurate name
zZkpUTXOStatus. Additionally, React components are updated to use the
updated totalSelector.

Issue-URL: https://app.clickup.com/t/363b42n