syncEnv.ts: compare controller times with Staking terms
Rather than comparing AdvancedStakeRewardController values for
REWARDING_START and REWARDING_END with values from the local .env file
for ADVANCED_STAKING_T_START and ADVANCED_STAKING_T_UNLOCK
respectively, compare them with the allowedSince and lockedTill values
from the Staking terms.

This means that the local .env file does not need to have times
in sync for the check to be valid (although it does need to point to
the right staking and pool contracts at least).