Clean up Faucet contract Without this patch, the `Faucet` contract is not as clean as expected.
This patch solves the problem by improving the variable names and deleting the unnecessary logic.
The Below variables/functions have been renamed:
`isAllowedToDrink` => `withinDrinkLimit`

`toggleRestrictToWhitelisted` => `updateRestrictToWhitelisted`

`addWhitelistedMultiple` => `whitelistBatch`

The `restrictToMaxDrinkCount` variable and `toggleRestrictToMaxDrinkCount` have been removed. We simply set `maxDrinkCount` to 0 for unlimited drink count.

The `claimNative` and `claimErc20` functions have been replaced with `withdraw`. this function accepts the `_claimedToken` param which is a token address to be withdrawn. If this address was zero address, it means the native token is going to be withdrawn.

Issue-URL: https://app.clickup.com/t/2kpham4