dapp: Extract back button from modal dialog
Without this patch, we have to redefine the BackButton every time we need to use it.

This is a problem because it breaks the reusability principle in UI design, and causes a lot of DRY violations.

This patch solves the problem by Extracting the BackButton from the modal dialog into a separate component.

Issue-URL: https://app.clickup.com/t/2htdyef