dapp: add progressToNewWalletAction
If a wallet action succeeded and we immediately want to initiate
another, rather than dispatching registerWalletActionSuccess followed
by startWalletAction, allow atomically progressing from the old action
to the new.  This eliminates a dispatch/re-render cycle and avoids
some race conditions where the re-render from the second dispatch
would conflict with the first if the dispatch was triggered via
useEffect.

A more general solution would be to use an existing npm module for
batching actions:

https://redux.js.org/faq/performance#how-can-i-reduce-the-number-of-store-update-events

but this simple approach is enough for now.