dapp: fix first scan of undefined UTXOs
The deps array of useEffect on PrivateBalance was deliberately empty
to avoid multiple automatic scans for undefined UTXOs when loading the
zAssets page.

However, this means that if we load the zAssets page while the wallet
is not connected, the useEffect() will fail to initiate the scan of
undefined UTXOs, since it would only run once on first page load, and
not again after wallet connection.

So add a variable to track whether UTXO scanning is needed / in
progress / complete, and use it to ensure that the scan is done just
once at the right time.