Fix too long display time of info messages
Without this patch, we have an annoying user experience when an info alert pops up because it takes for too long.

This is a problem because user might not like seeing multiple alerts which are not going to fade out quickly.

This patch solves the problem by changing the duration param in multiple parts of the codebase where the openNotification helper is called for info messages.

Issue-URL: https://app.clickup.com/t/36u8q9w