dapp: store stakeAmount value in redux
Without this patch, the stake amount input is not handled correctly,
it was stored in StakingInput component state.

This is a problem because we need to pass it into some children components
where we  need to use it.

This patch solves the problem by creating stakeAmount slice in redux to store this value.

Issue-URL: https://app.clickup.com/t/2cdtrj2

- [x] The commit messages all explain WHY the changes are needed.
- [x] The commit messages all explain HOW the changes solve the problem.
- [x] The commits are all strongly related.
- [x] None of the commits introduce any breakage.
- [x] None of the commits introduce any new technical debt.
- [ ] The commits update any related documentation.
- [ ] The commits include automated testing.
- [x] The commits have been thoroughly tested.