Make all decimals signs equal in Staking page.
Without this patch, we have some tiny inconsistency in the format of decimals in the Staking page.

This is a problem because we better abide by the same rule for decimal signs all across the codebase.

This patch solves the problem by simply using a unified pattern in those inconsistent cases.

Issue-URL: https://app.clickup.com/t/36u8h73