zAssets page - correct design of "closed data" view making it equal to a detailed view
Without this patch, we have a simple abnormality in terms of our zAssets page table width behavior when the data table is closed (not showing).

This is a problem because the width of the zAssests wrapper component shrinks when we click to hide the table info.

This patch solves the problem by changing the max-width css value from "max-content" to a fixed width (70rem) in zAssets style.scss file.

Issue-URL: https://app.clickup.com/t/36u8mvt