Set NFT rewards limit value by the owner
Without this patch, The Controller needs to hold to NFTs that are going to be allocated as rewards to stakers

This is a problem because the NFTs should be minted to the contract manually. It requires sending multiple transactions because we may not be able to mint all the NFTs in one single tx.

This patch solves the problem by adding the `setNftRewardLimit()` that can be executed only by the owner. this function accepts `_desiredNftRewards` as a parameter to define how many NFTs can be minted (allocated) as rewards by the controller.

The `prepareRewardsLimit` has also been changed to `updateZkpAndPrpRewardsLimit` because it does not prepare all the rewards anymore. It only prepares the $ZKP and $PRP rewards.

The `safeGetUnusedGrantAmount` also has been removed from `TransferHelper`. We use `IPrpGrantor` to access the `getUnusedGrantAmount()`