dapp:add comma in stake button amount
Without this patch, we have a tiny inconsistency in
how we display the stake amount to the user in the stake btn.

This is a problem because we have used a comma in our
balance card to show the zkp balance, but not used it in the stake button.

This patch solves the problem by using the
formatCurrency helper function in StakingBTN componenent.

Issue-URL: https://app.clickup.com/t/2zfk33c