dapp: Show precise values using tooltip
Without this patch, the staked amount and rewards are rounded
incorrectly, which give the user an impression of losing some tiny
amounts of value.

This is a problem because it may confuse the user.

This patch solves the problem by adding a tooltip to show the full
precision value of the rounded one.

Issue-URL: https://app.clickup.com/t/2yymmap