dapp: extract sumBigNumbers() into @crypto workspace
Without this patch, we have a violation of separation of workspaces, bcz
it is wrong for anything in dapp/ to depend on anything in contracts/ folder.

This is a problem because now it breaks hardhat tasks execution as it
tries to compile dapp workspace.

This patch solves the problem by moving sumBigNumbers() from dapp folder
into @crypto workspace.

Issue-URL: https://app.clickup.com/t/35tbwcv