dapp: change of the utxoData containing 3 UTXOs
Without this patch, the dApp works with utxoData containing only 2
UTXOs. More specifically it has a check for the length of the utxoData.

This is a problem because the new version of smart contracts has 3 UTXOs.

This patch solves the problem by changing the check for the length of
the utxoData to the length of the utxoData containing 3 UTXOs. Also, the
patch adds logging of the error in the console.