dapp: add goerli testnetwork
Without this patch, we are not able to stake in dApp on the Goerli
network.

This is a problem because the Goerli network is the network where we can
test the bridging between Ethereum and Polygon.

This patch solves the problem by adding the Goerli testnetwork in dApp
via adding connectors and required contract env variables. Also,
amplify.yaml is updated.

Issue-URL: https://app.clickup.com/t/35krat2