dapp: update balance card
Without this patch, the balance card shows 0 zZKP and 0 PRP rewards on
the Goerli network. The Goerli network now allows advanced staking but
cannot fetch the generated UTXOs yet.

This is a problem because if the network hasAdvancedStaking flag true,
the balance card shows zZKP and PRP rewards balances. In the case of the
Goerli network, they are zero because they are not fetched from the
subgraph.

This patch solves the problem by adding the utility function
chainHasPoolContract(chainId). If the selected chain has the Pool
contract then the rewards and UTXOs can be fetched and the balance can
show zZKP and PRP rewards.

Issue-URL: https://app.clickup.com/t/35krat2