dapp: add message about bridging on staking info panel
Without this patch, the user would not know that bridging and creating
of the UTXOs can take up to 10 minutes.

This is a problem because it may result in a bad UX and frustraction of
the users.

This patch solves the problem by adding a message to the staking info
panel if the stake is made on the Ethereum network.