dapp: hook up total staked component
Without this patch, we have not synced data in TotalStakedCard &
ClaimedProgress components.

This is a problem because our dapp UI must be showing real & up-to-date
data.

This patch solves the problem by adding a new getter function for
AdvancedStakeRewardControllerContract that uses the chainId of the MASP.
This change allows calling the contract and getting the data on any
chain. The data includes totals(), limits() public structs of
AdvancedStakeRewardControllerContract. These structs contain values of
total and claimed rewards to update ClaimedProgress component as well as
scZkpStaked of scaled staked ZKP to update TotalStaked card. The
totalStaked value is saved as a BigNumber string in the Redux store.
Also, ClaimedProgress is extracted as a separate component and
syncEnv.ts file was updated to work with a new getter function of
AdvancedStakeRewardControllerContract.

Issue-URL: https://app.clickup.com/t/2xpbr05