dapp: update staking and unstaking text
Without this patch, the text about the advanced staking lock period was
tightly coupled to the single unstake date. However, with the
introduction of the MASP v0.5 each advanced stake has its lock period.

This patch solves the problem by changing the text about the lock period
of the stakes and showing the period for which the stakes are locked.

Issue-URL: https://app.clickup.com/t/36u93b1