dapp: fix remaining days banner
Without this patch, when the staking is not yet open, the remaining days
banner shows the remaining days till the staking is closed.

This is a problem because it may confuse a user that staking is not
working.

This patch solves the problem by showing the remaining days till the
staking is open if it has not yet started.

Issue-URL: https://app.clickup.com/t/364w4ga