dapp:  disable total unclaimed rewards in the unstake tab
Without this patch, the total unclaimed rewards shows the rewards based
on UTXO in the pool. This confuces a user because the stake rows shows
rewards and the total of stakes do not match that is in the pool. This
happens when some the stakes or rewards are claimed.

This patch solves the problem by removing the total rewards estimation
in the unstake tab.

Issue-URL: https://app.clickup.com/t/394myz5