dapp: fix the math formula for the rewards calculation
Without this patch, the rewards were calculated incorrectly. The
calculation assumed that T_END of the stake is always the same unlocking
date for all stakes. This is not the case. In v0.5 of smart contracts,
the unlocking date of the stake is unique for each stake (e.g., 90 days
after the stake). As a result, the rewards are calculated incorrectly.

This patch solves the problem by calculating the rewards based on the
lockedTill date of the stake. This patch also updates the calculation of
the APY to use ADVANCED_STAKING_T_END instead of T_UNLOCK. Updated
tests.

Issue-URL: https://app.clickup.com/t/394n6ne