dapp: get the rewards from the subgraph, otherwise use the math formula
Without this patch, the rewards for the existing stakes were calculated
based on the math formula.

This is a problem because it gives less than 1% of inaccuracy which is
enough to cause the confusion.

This patch solves the problem by fetching the rewards from the subgraph.
And if the reward is present in the subgraph, then the math formula is
replaced with the rewards from the subgraph. This patch also changes the
error handling of the request to the graph. Instead of throwing the
error, it returns it for graceful handling.

Issue-URL: https://app.clickup.com/t/36g9pkh