dapp: add totalsOfAdvancedStakes slice to fetch the data just once
Without this patch, the components fetched the data directly from the
chain.

This is a problem because the data is fetched on every render.

This patch solves the problem by creating a redux slice with aggregated
data about advanced staking progress. The slice contains the totals of
1) staked ZKP,  2) vested zZKP rewards, and 3) claimed zZKP rewards. The
slice is then used to fetch the data just once and render it in the UI
where needed.